+++ CHANGELOG.md
@@ -27,6 +27,8 @@ All user visible changes to this project will be documented in this file. This p
     - Redesigned auth page. ([#564], [#533])
     - Redesigned login modal. ([#564], [#533])
     - Redesigned language selection modal. ([#533])
+    - Style page:
+        - Redesigned colors tab. ([#616], [#614])
 
 ### Fixed
 
@@ -52,6 +54,8 @@ All user visible changes to this project will be documented in this file. This p
 [#599]: /../../pull/599
 [#604]: /../../pull/604
 [#609]: /../../pull/609
+[#614]: /../../issues/614
+[#616]: /../../pull/616